backend/requirements.txt
# É FESTA — Dependências Python
# Instale com: pip install -r requirements.txt

# ── Framework Web ──────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0

# ── Banco de Dados (SQLAlchemy ORM) ───────────────────────────────────────────
sqlalchemy==2.0.30
alembic==1.13.1              # Migrações de banco de dados

# ── Drivers de Banco ──────────────────────────────────────────────────────────
# SQLite (embutido no Python — sem instalação extra)
# Para PostgreSQL em produção, descomente:
# psycopg2-binary==2.9.9

# ── Autenticação & Segurança ──────────────────────────────────────────────────
python-jose[cryptography]==3.3.0   # JWT (tokens de acesso)
passlib[bcrypt]==1.7.4             # Hash de senhas (bcrypt)
python-multipart==0.0.9            # Upload de arquivos

# ── Mercado Pago ──────────────────────────────────────────────────────────────
mercadopago==2.2.2

# ── Configurações ─────────────────────────────────────────────────────────────
pydantic-settings==2.2.1
python-dotenv==1.0.1

# ── HTTP Client (Google OAuth) ────────────────────────────────────────────────
httpx==0.27.0

# ── Templates Jinja2 ──────────────────────────────────────────────────────────
jinja2==3.1.4

# ── E-mail ────────────────────────────────────────────────────────────────────
# smtplib é embutido no Python — sem instalação extra

# ── Validação de dados ────────────────────────────────────────────────────────
pydantic==2.7.1

# ── Utilitários ───────────────────────────────────────────────────────────────
python-slugify==8.0.4
Pillow==10.3.0               # Redimensionamento de imagens

# ── Desenvolvimento ───────────────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0                # Cliente HTTP para testes
